Handle Multi-IP Address Systems
In cases where the system is a VM, or has multiple NICs, it can have
many internal IP Addressess associated with it. This is often done in
systems that carry multiple services: MX, Web Server, and others all
running on a single system, or a single system running multiple websites
at different addresses. This code change covers that scenario, as well
as strictly defining variable types.